+++ ch6-entitymanager/src/main/java/ch6/p1_container_managed/transaction_scoped/service/DepartmentService.java
@@ -33,7 +33,7 @@ public class DepartmentService {
     }
 
     @Transactional // creates transaction
-    public void assignEmployee(int emplId, int projId) {
+    public void assignEmployee(long emplId, int projId) {
         Employee employee = em.find(Employee.class, emplId);
         Department department = em.find(Department.class, projId);
         department.getEmployees().add(employee);
