+++ ch6-entitymanager/src/main/java/ch6/p1_container_managed/transaction_scoped/service/EmployeeService.java
@@ -40,7 +40,7 @@ public class EmployeeService {
     }
 
     // transaction is active, uses this em with active persistence context
-    private void logEmployee(final int id) {
+    private void logEmployee(final long id) {
         System.out.println("Same EntityManager: " + em.find(Employee.class, id));
     }
 }
